wsdiscovery only allows one ip address
