+++ src/sdc11073/wsdiscovery/__init__.py
@@ -1,3 +1,14 @@
-from .wsdimpl import WSDiscoveryBlacklist, WSDiscoveryWhitelist, WSDiscoverySingleAdapter
-from .wsdbase import MATCH_BY_LDAP, MATCH_BY_UUID, MATCH_BY_STRCMP, MATCH_BY_URI, match_scope
+# noqa: D104
 from .common import MULTICAST_PORT
+from .wsdimpl import (
+    MATCH_BY_LDAP,
+    MATCH_BY_STRCMP,
+    MATCH_BY_URI,
+    MATCH_BY_UUID,
+    WSDiscovery,
+    WSDiscoverySingleAdapter,
+    match_scope,
+)
+
+__all__ = ['MULTICAST_PORT', 'MATCH_BY_LDAP', 'MATCH_BY_UUID', 'MATCH_BY_STRCMP', 'MATCH_BY_URI', 'match_scope',
+           'WSDiscovery', 'WSDiscoverySingleAdapter']
