+++ tests/test_compression.py
@@ -5,7 +5,7 @@ from lxml import etree
 from sdc11073.sdcclient import SdcClient
 from sdc11073.xml_types.pm_types import InstanceIdentifier
 from sdc11073.httpserver import compression
-from sdc11073.wsdiscovery import WSDiscoveryWhitelist
+from sdc11073.wsdiscovery import WSDiscovery
 from sdc11073.location import SdcLocation
 
 XML_REQ = '<?xml version=\'1.0\' encoding=\'UTF-8\'?> \
@@ -27,7 +27,7 @@ class Test_Compression(unittest.TestCase):
 
     def setUp(self):
         # Start discovery
-        self.wsd = WSDiscoveryWhitelist(['127.0.0.1'])
+        self.wsd = WSDiscovery('127.0.0.1')
         self.wsd.start()
         # Create a new device
         self.location = SdcLocation(fac='tklx', poc='CU1', bed='Bed')
